Rename SerializedFrame enums
Summary:
SerializedFrame is very similar to `std::borrow::Cow<'a, [u8]>`, except that its owned type is Bytes instead of Vec. As a result we can't use Cow for our purpose. Byte is similar to Vec instead of it's ref counted, which is easier to use in our case.

Rename SerializedFrame enum from Bytes to Owned and Slice to Borrowed to clarify ownship model.

Reviewed By: boyuni

Differential Revision: D34061719

fbshipit-source-id: b86c019dd5e0c1ae50d6004ef787f1a618d20492