Adding snapshot command
Summary:
Adding a new command `snapshot` to the CLI utility.

Usage:
```
below snapshot [OPTIONS] --begin <begin> --end <end> [host]
```

Currently when a user wants to create a dump (snapshot) in order to replay it later, they need to do it in a more tedius way involving multiple calls to be able to create the files required to replay such as below:
```
$ below debug convert-store --start-time <start-time> --end-time <end-time> --to-store-dir <snapshot-dir>
# write a custom below.conf with custom store_dir
$ below --config <my-config-with-snapshot-dir> replay
```

This new command will essentially simplify this by performing the action to call `debug` and store all the files needed to replay in a tarball. This tarball can then be extracted and used to reply.

Reviewed By: brianc118

Differential Revision: D34161168

fbshipit-source-id: 2c5146279f87c1f4ebed150aee796cfc4706b31a