Implement collector plugin framework
Summary:
Add a collector plugin framework that allows new collectors to be
easily added.

For example, consider an expensive collector that takes seconds to
gather data. We can simply implement a read function for this
collector, run the collector plugin on a separate thread, and read
samples on the main collection thread.

**Latest sample**

Only the latest sample is saved at any time. If the consumer thread is
asking for samples more frequently than the collector thread is
collecting, it is up to the consumer thread to cache the previously
consumer sample and determine if it should continue to be used (e.g.
based on timestamp of previous sample if this is provided by the
collector).

**Error handling**

We consider two types of errors at collection. Recoverable errors are
intended to be handled by the collection thread and are invisible to
the consumer thread. Unrecoverable errors are handled by the
collection thread and also sent to the consumer thread.

Reviewed By: dschatzberg

Differential Revision: D34192301

fbshipit-source-id: d823fcc1d18eb4863ac640716ba8d707b96af5a2