Remove slog_async
Summary:
slog_async can result in us not outputting log messages on exit. The
core issue is that we pass our logger around including to other threads which
might not exit until after the main thread exits.

This issue was reported on Github:
https://github.com/facebookincubator/below/issues/8156

This diff changes the logic to just synchronously log.

slog_async gives us some confidence that our main thread won't be blocked if
logging is slow. I'm not sure how much this is an issue in practice, but it's
something we should keep an eye out for.

I see a couple alternatives to this approach:

1) We continue using slog_async, but just always ensure we tear down threads
explicitly when the main process exits. This would probably require our own
ThreadBuilder to make this an easy pattern to enforce.

2) We continue using slog_async, but rather than passing a logger all around we
have a single global logger that we can explicitly destroy when main() exits.

Reviewed By: lnyng

Differential Revision: D38618612

fbshipit-source-id: 3e20d95f0e50e83974e7524effc5cf033ec00b9c