Filter by other fields (Part 1)
Summary:
## Task Overview (both part 1 & part 2)
Support a filter by selected column/each field instead of filtering by name. A view can have multiple tabs and each tab has multiple fields. For example process view has "General" "CPU" "MEM" and "I/O" tabs which all display the same rows but different fields for the model depending on which tab you are on.

A filter can be applied to only one column/field at a time and it is enabled for string columns (not numeric yet).  A future improvement would be to enable filtering for multiple columns at once and allowing filtering on numeric columns (>, <, and = ).

## Part 1 of Task
With this current feature, `filter` and `filter_cache_for_zoom` each hold a pair that maps `FieldId -> String`. In the future, to support multiple filters, we might replace this with a map of all fields -> filter strings.
This pair will be correctly updated with the selected FieldId and the filter string whenever the user filters. Currently, there should be no change to the filtering behavior from the user's pov. Filtering the rows based on FieldId will be introduced in part 2

## Context

See https://github.com/facebookincubator/below/issues/8115

The desired behavior is this: if we filter by a particular field only present in a one of the tabs, we will keep the filter even when we switch tabs.
For example:
1. Enter process view on general tab
2. Select "State" column (press '.' a few times)
3. Filter with "IDLE" state
4. Enter CPU view (<TAB>)
5. We should observe the same processes/rows are displayed as we keep holding <TAB> to move across the process view tabs.

Reviewed By: brianc118

Differential Revision: D38760054

fbshipit-source-id: 2cf50b74559290a3bcf5019a1d0954c6f5fb764b