Filter by other fields (Part 2)
Summary:
## Task Overview (both part 1 & part 2)
Support a filter by selected column/each field instead of filtering by name. A view can have multiple tabs and each tab has multiple fields. For example process view has "General" "CPU" "MEM" and "I/O" tabs which all display the same rows but different fields for the model depending on which tab you are on.
A filter can be applied to only one column/field at a time and it is enabled for string columns (not numeric yet).  A future improvement would be to enable filtering for multiple columns at once and allowing filtering on numeric columns (>, <, and = ).

## Part 2 of Task
This diff introduces the ability to filter on any column. To filter on a specific column, select a column by pressing '.' or ',' and then filter
- command palette should list both the filter column and filter string
- filtering on columns should work with all the views
- the command palette is updated when the filter is updated or the view is changed
- filter popup should say "Filter by {selected col name}".

## Context

See https://github.com/facebookincubator/below/issues/8115

The desired behavior is this: if we filter by a particular field only present in a one of the tabs, we will keep the filter even when we switch tabs.
For example:
1. Enter process view on general tab
2. Select "State" column (press '.' a few times)
3. Filter with "IDLE" state
4. Enter CPU view (<TAB>)
5. We should observe the same processes/rows are displayed as we keep holding <TAB> to move across the process view tabs.

Reviewed By: brianc118

Differential Revision: D38835737

fbshipit-source-id: fafff81c16f1c8a1295d63b6bda9bed7e1c4daae