Allow snapshot files to be renamed/moved
Summary:
We have a problem with the snapshot tar's that when a user renames them they are unable to be used since the folder inside that contains the original name of the snapshot still exists when unarchiving. Same issue occurs when moving as we may not be able to preserve the exact path.

E.g

Lets assume the snapshot was created as

```
snapshot_01666902362_01666902482.4RSwEMvVkx8c
```

The tar will still contain
```
/tmp/snapshot_replay.XxXX/snapshot_01666902362_01666902482.4RSwEMvVkx8c
```

But since we use the new name of the file to iterate, the replay command fails as it is unable to find

```
/tmp/snapshot_replay.XxXX/<some_new_name>
```

By retriving the name of the original snapshot folder from the unarchived data we will be able to use the original name avoiding any issue with renaming or moving the tarball.

Reviewed By: brianc118

Differential Revision: D40773235

fbshipit-source-id: 3996d6556e9a6b0ede89b766ba1c1c20a9dbc82f