Document `cgroup_root` config option
This is a useful config to document b/c many systems are still running
under the (rather weird) hybird cgroup1 and cgroup2 system (ie.
`/sys/fs/cgroup/unified` for cgroup2 and `/sys/fs/cgroup` for cgroup1).

On hybrid systems, it is useful to set `cgroup_root` to the unified
mountpoint as a stopgap to make below run.