Document `cgroup_root` config option (#8179)
Summary:
This is a useful config to document b/c many systems are still running under the (rather weird) hybird cgroup1 and cgroup2 system (ie. `/sys/fs/cgroup/unified` for cgroup2 and `/sys/fs/cgroup` for cgroup1).

On hybrid systems, it is useful to set `cgroup_root` to the unified mountpoint as a stopgap to make below run.

Pull Request resolved: https://github.com/facebookincubator/below/pull/8179

Reviewed By: lnyng

Differential Revision: D43249631

Pulled By: brianc118

fbshipit-source-id: 838851e30760500788ec96d436967acc3118a5a3