Add NStgid column to process view
Summary:
For processes running inside pid namespace, they often see a different self pid than from below's perspective. If those namespaced pids are logged, it could be hard to correlate with below data.

This diff adds support to display the NStgid info obtained from /proc/<pid>/status. It contains a list of pids of a process that belongs to all of its pid namespaces. Since the first entry is always the same as the existing pid column, it's skipped to save some visual space.

Reviewed By: brianc118

Differential Revision: D43845699

fbshipit-source-id: be24224174ca1833809d24a3897fcdf9185809c8