command palette shows selected field name and raw value
Summary:
Command palette is the bottom bar of the below UI, which is used to display alerts, information about current selected entry, and for inputting vim-style commands.
Info for selected entry is the most commonly used. It currently shows the selected entry's full cgroup path on cgroup view, and full command line in process view. It does not change on column selection.
This diff adds support to make the command palette show column-related info, like `<field name> : <raw value>`.
For both cgroup and process view, when the first column is selected, they still show full cgroup path and full command line, same as the previous behavior.
When different columns are selected, the command palette will be updated with the raw value of the selected cell, like `cpu.usage_pct : 422.79345199873654`
The field name can be used in `below dump`. The raw value addresses the problem that the column sometimes maybe too narrow to show the whole value or if the value is not precise enough.
Other views do not have support yet.

Reviewed By: brianc118

Differential Revision: D43935299

fbshipit-source-id: 4704d7c3d038953308436931d10ea36eb9092a60