snapshot: Support specifying snapshot output file (#8181)
Summary:
Previously, the snapshot file was always generated in current working
directory. Sometimes it is useful to specify where the snapshot file
should be place and what name it should have. It saves a couple of
steps, especially when the snapshot is intended to be exported off a
remote machine.

Pull Request resolved: https://github.com/facebookincubator/below/pull/8181

Reviewed By: dschatzberg

Differential Revision: D44026824

Pulled By: brianc118

fbshipit-source-id: 0122ad1b04d21594a0c14ada454b6cbcfc7e6fc7