Report zswap and zswapped from memory.stat
Summary: memory.stat has zswapp and zswapped field which measure the compressed and decompressed size of the zswapped pages, respectively. zswap field is the same as memory.zswap.current and should replace the latter to one read. zswapped is useful to understand how much swap is on disk vs compressed. memory.numa_stat does not contain the two fields.

Reviewed By: antonis-m

Differential Revision: D49345444

fbshipit-source-id: 3858fb9a6436d94072ca4da62f3f990d846e6973