Revert "Deprecate Stat::cpus field" (#11)
This reverts commit eba5bc32f0d61c5a145819f9271dcd62189535a4.

We are still running older versions of below in production that still
write to `cpus`. Upstream has deprecated support for the old `cpus`
field in favor of `cpus_map` which is basically identical, except that
it suports offlined cpus.

We still need our clients to be able to support older data. So revert
the upstream commit until we have gotten all our customers onto newer
releases. Reverting allows us to access percpu data collected in the field.
Without the revert, we "lose" that data. Technically it's still in the data store,
but the client just cannot parse it.

Co-authored-by: Daniel Xu <dxu@dxuuu.xyz>