Fix build with no-vendor feature
Summary:
The build was broken because as of D53102023, libbpf-rs no longer provided the
novendor feature, as it was replaced with the proper set of default features.

With this change we explicitly enable said default features by this program's
default feature.

Furthermore, to make no-vendor do the right thing, require that it be used
*without* default features via the build script. This is admittedly a breaking
change, but it seems to be unavoidable: there is no sensible way to combine
negative features ("no<...>") with positive ones. Cargo strongly suggests that
all features be additive in nature, and no-vendor violates that. However, that
can't be easily fixed, because of zstd's feature set.

Reviewed By: dtolnay

Differential Revision: D53323353

fbshipit-source-id: b36a7a84892e5b6ec2a4888649b6c737148d6a9b