below: ioctl: Ensure ioctl(SIOCETHTOOL) compat across libc implementations
Summary:
Different libc implementations (e.g., musl and glibc) define the `ioctl`
request argument with different types (`c_int` for musl and `c_ulong` for
glibc). In C, this difference is masked by typeless #defines, but in Rust, it
causes a type mismatch that must be explicitly handled:

    error[E0308]: mismatched types
       --> below/ethtool/src/reader.rs:36:58
        |
    36  |     let exit_code = unsafe { libc::ioctl(fd.as_raw_fd(), nix::libc::SIOCETHTOOL, &ifr) };
        |                              -----------                 ^^^^^^^^^^^^^^^^^^^^^^ expected `i32`, found `u64`
        |                              |
        |                              arguments to this function are incorrect

To ensure portability and prevent type errors, we now convert
`libc::SIOCETHTOOL` to the appropriate type. This conversion ensures that our
code remains compatible with both musl and glibc.

Reviewed By: dschatzberg

Differential Revision: D58168265

fbshipit-source-id: d19789150d73dbe8fb0351cdfd4a02d79ea01400