below: cmdline: Avoid MPSC queue overhead in priority inversion avoidance
Summary:
mpsc channels manage a queue of messages that they receive from multiple
producers. This queue management involves additional synchronisation
overhead to ensure that messages are correctly enqueued and dequeued
with proper consistency. In our case, we only have one producer and one
consumer, so the complexity and overhead of managing multiple producers
is unnecessary, and we can just rely on a cheap Mutex for
synchronisation.

This saves about 0.5-1% of CPU instructions per iteration due to the
reduction in synchronization overhead associated with the "m" part of
the mpsc channel.

Reviewed By: brianc118

Differential Revision: D58316866

fbshipit-source-id: 2cdbe78f845f68b3f8547abef7e22ef5eee88d8e