Store slabinfo in vec
Summary:
Looks like /proc/slabinfo may contain rows with same name. Therefore, it may not be the best idea to store the rows in a map keyed by the name column. Just store the slabinfo entries in a vec instead and preserve the order.
Update the slabinfo field name in the recorded sample so that we don't try to deserialize and map into a vec. Previously recorded slabinfo data will be ignored.

Reviewed By: hnaz

Differential Revision: D66518742

fbshipit-source-id: d30423b91c48410011c9f53ad3195e145947c8f8