Update `libbpf-rs` & `libbpf-cargo` to `0.25.0-beta.1`
Summary:
Update libbpf-rs and libbpf-cargo to 0.25.0-beta.1. One of the main
changes for this release is the adjustment to the way libbpf-cargo
exposes C enums in Rust code: earlier we were creating Rust enums, but
because these have certain inherent limitations, we have switched over
to using newtype style structs and named public constants for
representing variants.

Reviewed By: slawlor

Differential Revision: D67880323

fbshipit-source-id: e626de5f2a3cf79a67d2898cf932198901bb104e