below: proc_reader: Skip cmdline task enqueue when threadpool is saturated
Summary:
When reading /proc/<pid>/cmdline, we now check if the threadpool is fully
occupied (i.e., the number of active workers is equal to or exceeds the maximum
allowed). This avoids a useless stall for 20ms waiting for a threadpool worker
to become available.

This check is racy -- we may erroneously determine that the pool is full when
it isn't -- but that risk is acceptable since it can only lead to an early exit
rather than blocking work that should be executed. In the typical scenario, the
threadpool is completely idle when we come here.

Reviewed By: dschatzberg

Differential Revision: D70254923

fbshipit-source-id: 027e0ef82321079c8786fe9c389dae240f38dacc