Move to column major layout for summary view
Summary:
Before, it was row major, meaning every "grouping" was spread out on a single row. This made visually scanning grouped data had to parse - think of a spreadsheet but without the ability to select a row or column.

This diff moves the entire section to column major. This means data within the same group is organized vertically in a single column. This not only conserved screen real estate but also makes it easier to visually parse.

Reviewed By: lnyng

Differential Revision: D74614390

fbshipit-source-id: 1374ad669e271ff09da94fed4790d8996b06eed0