Switch from bold+underline on tabs to indicator
Summary:
Before, selected tabs were drawn with bold+underline. My feeling is that it's visually suboptimal.

This diff replaces it with a visual indicator. The advantage with this inline indicator is that when you switch tabs it makes the text "move" a little. Which draws the eye and provides a more satisfying visual response. I like this more than highlighting.

Reviewed By: boyuni

Differential Revision: D74751812

fbshipit-source-id: 418fb44e02c564c082afca8ccc15d28678af7621