Log to $TMPDIR by default
Summary: Currently interactive modes log to temporary file if /var/log/below.log cannot be opened. This happens if below runs as non-root which is very common. This diff makes below log to /tmp/below.log by default which is unlikely to fail. This will not affect record mode as it logs to stderr and relies on journald/rsyslog for log persistence and rotation.

Reviewed By: danobi

Differential Revision: D74909065

fbshipit-source-id: 60f8432ffd65607456a34e2b9617af7a6f550d83