Collector plugin do not send error to consumer
Summary:
Below has a collector plugin framework to collect extra data async. Currently if the collector fails (timeout, incorrect config, dependency missing etc), an error is returned to the collecting thread, and also stored as a shared value later retrieved by the consumer (main) thread. When the consumer sees the error, the whole sample fails.

This is undesirable because we already log the error in the collecting thread. Below follows the model that most failures result in a None field of the sample instead of failing the whole sample.

This diff updates to log the error in the collecting thread, and report None to the consumer for any error.

Differential Revision: D76748582

fbshipit-source-id: 96b095e215cedb8fbb176e6801a25751f047783b