refactor: Optimize `Dockerfile`
The switch to a Fedora base image is much more modern vs Ubuntu 22.04:
- Smaller list of deps to declare / maintain.
- Can use `cargo` package or `rustup` instead of relying on Rustup shell install method.
- Simple to support minimal runtime image + cross-compilation.
- Builder stage is 2.47GB (Ubuntu 22.04) vs 1GB (Fedora 42).
- Runtime image is 159MB (Ubuntu 22.04) vs 30MB (Fedora 42).

Swapped `--all-targets` for `--bin below` as this is the only artifact of interest for the final runtime image.

The `Dockerfile` HereDocs feature was used for easy to grok multi-line `RUN` instructions. The prior `Dockerfile` had multiple `RUN` layers that could be merged to avoid wasted disk usage.

`ENV LANG=C.UTF8` is not required anymore:
- Default for Fedora base image.
- The `cursive` crate dep was switched from the locale affected `ncurses` backend to the `crossterm` backend some time ago, this backend is also the default in Cursive since `0.21.0` (Aug 2024).