chore: `Dockerfile` replace `cargo` with `rustup`
Provides a bit more flexibility and better matches original `Dockerfile` usage of Rustup.

This comes at the cost of about 300MB extra as it bundles it's own LLVM and the `librustc_driver` library is almost twice the size of the equivalent from the `cargo` package.

Builder image increases to 1.36GB.