chore: Use Zig to support more glibc versions
This adds a bit more complexity to support building for a lower glibc target to better support the Debian packaging stage so that the built binary will be compatible with older versions of glibc.

The default without a version suffix is currently glibc 2.28 (Aug 2018) for Zig releases v12 to v14.1, but will likely be bumped to glibc 2.34 (Aug 2021) with the upcoming Zig v15 release.
- Ubuntu 18.04 uses glibc 2.27
- Ubuntu 20.04 uses glibc 2.31
- Ubuntu 22.04 uses glibc 2.35
- Debian 11 (Aug 2021) uses glibc 2.31
- Debian 12 (June 2023) uses glibc 2.36

`below` cannot build for glibc lower than 2.28, it relies on headers `libbpf_internal.h` from `libbpf-sys` crate that require the `fcntl64` glibc symbol (available since glibc 2.28). Which makes the default Ubuntu 18.04 for the since removed Debian `Dockerfile` invalid for builds.

Bringing in zig adds 500MB of weight as it linked to a different LLVM release (19 in this case, while `cargo` and current Rustup stable toolchains use LLVM 20 with `cargo` and `clang` packages sharing the `llvm-libs` package). Bulk of the `zig` package weight is:
- llvm-19 (288MB)
- clang-19 (52MB)
- zig-14 (146MB)

`cargo install cargo-zigbuild` adds approx 50MB during the build/install process that could be optionally removed `/root/.cargo/registry`.

With Rustup the image weight for the builder image grows to 1.9GB (or 1.6GB with `cargo` instead of `rustup` package).