Update cursive 0.20.0 -> 0.21.1 (#8253)
Summary:
Pull Request resolved: https://github.com/facebookincubator/below/pull/8253

This is an attempt to update cursive. What started this was a RUSTSEC on a transitive package but it became a bit of rabbit hole.

Anyway, this diff (please review, don't just stamp) updates cursive and it's (I think) only client, which is resctl.

This update comes with non-trivial breaking changes, in particular:

- The View now requires Send + Sync, to allow accessing or moving views between threads. This prevents using Rc/RefCell, and may require using Arc/Mutex instead. This should eventually open the way for more multi-threaded processing of the view tree.

Of course this requires changing pretty much everything that touches Views... which is a lot.

DevMate did a lot of the work once I got it in the right track. It builds and tests pass.

The other thing worth mentioning is that the crate cursive_buffered_backend is removed, because cursive now includes this according its CHANGELOG.

- The output to the backend is now buffered and delta-patched, resulting in improved performance for most backends.

Reviewed By: lnyng, Imxset21

Differential Revision: D77703952

fbshipit-source-id: f67cd3591cde1f86eceadbea25ddd061b599e0de