Rename RUSTFIXDEPS to RUSTUNUSEDDEPS
Summary:
The linter RUSTFIXDEPS only checks for one thing: unused dependencies. This is a frequent source of confusion: it's not autodeps, and never adds dependencies. We have RUSTGUESSDEPS for adding missing dependencies.

(The name RUSTFIXDEPS is because it consumes rustc error messages like RUSTFIX does.)

Rename the lint rule to RUSTUNUSEDDEPS to avoid confusing users.

Reviewed By: ezgicicek

Differential Revision: D80335007

fbshipit-source-id: 6310f2356816b3b1fbc75a1eba8595ce0386e5b2